Check Linux window system using $WAYLAND_DISPLAY
XDG_SESSION_TYPE is a systemd(1) environment variable and is unlikely
set in non-systemd environments. The WAYLAND_DISPLAY variable should
only be set when using the Wayland compositor. On the other hand, we
can't use the DISPLAY environment variable since many use XWAYLAND.
Therefore, we only check if the WAYLAND_DISPLAY environment variable is
set and fall back to x11.